API

API Reference

DeIdentification.DeIdDataFrame — Method.

DeIdDataFrame(df, hash_cols, salt_cols, dateshift_cols, drop_cols, dateshift_dict, id_col, id_dicts, salt)

This is the constructor for our DeIdDataFrame objects. Note that the id_col is the primary identifier for the dataset and what we use for our lookup in the date-shift dictionary. Also note that the dateshift_dict object stores the Research IDs as the keys, and number of days that the participant (for example) ought to have their dates shifted. The id_dicts argument is a dictionary containing the hash digest of original IDs to our new research IDs.

source

DeIdentification.DeIdentified — Method.

DeIdentified(cfg)

This is the constructor for the DeIdentified struct. We use this type to store arrays of DeIdDataFrame variables, while also keeping a common salt_dict and dateshift_dict between DeIdDataFrames. The salt_dict allows us to track what salt was used on what cleartext. This is only necessary in the case of doing re-identification. The id_dicts argument is a dictionary containing other dictionaries that store the hash digest of the original primary IDs to our new research IDs.

source

DeIdentification.deidentify — Method.

deidentify(config_path)

Run entire pipelint: Processes configuration YAML file, de-identifies the data, and writes the data to disk. Returns the DeIdentified object.

source

Base.write — Method.

write(deid)

Writes DeIdentified structure to file. The datasets are outputted as CSVs, the dictionaries are written to josn. The files are written to the output_path specified in the configuration YAML.

source

DeIdentification.dateshift_col! — Function.

dateshift_col!(df, date_col, id_col, dateshift_dict)

This function is used internally by dateshift_all_cols!(). We require that date shifting is done at the patient level. Thus, we pass the id_col to ensure that for a given patient, all their encounter data are shifted by the same n_days.

Arguments

df::DataFrame: The dataframe with the column to be date shifted

date_col::Symbol: Column with dates to be shifted

id_col::Symbol: Column with ID (e.g., patient ID) to find dateshift value for this observation

dateshift_dict::Dict: Dictionary where keys are ID (e.g., patient ID) and values are integers by which to shift the date (i.e., a number of days)

max_days::Int: The maximum number of days (positive or negative) that a date could be shifted

source

DeIdentification.hash_all_columns! — Method.

hash_all_columns!(df, salt_dict, hash_col_names, salt_col_names)

This function simply iterates over columns to be hashed, those to be salted and hashed, and those for which we want to generate sane-looking research IDs. Note that the id_dicts nested dictionary stores dictionaries for each column for which we have generated research IDs. Also note that this approach presumes if we have a column (e.g., :ssn) for which we generate a research ID in one dataframe, then all subsequent columns in other dataframes must have the same column name (i.e., :ssn).

Arguments

df::DataFrame: Dataframe to be de-identified

logger::Memento.Logger: This is our logger that writes logs to disk

hash_col_names::Array{Symbol,1}: Array with names of columns to be hashed

salt_col_names::Array{Symbol,1}: Array with names of columns to be salted and hashed

id_cols::Array{Symbol,1}: Array with names of columns to be turned in to research IDs

id_dicts::Dict{Symbol, Dict{String, Int}}: This is a dictionary of dictionaries. The keys of the outer dictionary are the column names of ID variables. The values of the outer dictionary are dictionaries themselves with key-values being hash digest IDs => research IDs.

salt_dict::Dict{String, Tuple{String, Symbol}}: Dictionary where key is cleartext, and value is a Tuple with {salt, column name}

source

DeIdentification.hash_column! — Method.

hash_column!(df, col_name, salt_dict)

This function is used to hash columns containing identifiers. Hashing is done in place using SHA256 and a 64-bit salt. Of note is that missing values are left missing.

Arguments

df::DataFrame: The dataframe to be de-idenfied

col_name::Symbol: Name of column to de-idenfy

source

DeIdentification.hash_salt_column! — Method.

hash_salt_column!(df, col_name, salt_dict)

This function is used to salt and hash columns containing unique identifiers. Hashing is done in place using SHA256 and a 64-bit salt. Of note is that missing values are left missing.

Arguments

df::DataFrame: The dataframe to be de-idenfied

col_name::Symbol: Name of column to de-idenfy

salt_dict::Dict{String, Tuple{String, Symbol}}: Dictionary where key is cleartext, and value is a Tuple with {salt, column name}

source

DeIdentification.rid_generation — Method.

rid_generation(df, col_name, id_dict)

This function is for generating a research ID from a hash digest ID. That is, for those cases in which we want to hash an ID (not salt) and preserve linkages across data sets, it's nicer to have a numeric value rather than 64-character hash digest to look at. Note that using this function pre-supposes that the rows of the data frame have been shuffled, as we do in the DeIdDataFrame() construnctor. Otherwise, we risk potentially leaking some information related to the order of the observations in the dataframes.

Arguments

df::DataFrame: The de-identified dataframe

col_name::Symbol: The column containing the hash digest of the original ID

id_dict::Dict: Dictionary with hash digest => research ID mapping

source